+++ packages/web-app/src/app/vscode/VsCodeBody.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { ACCENT_COLOR } from '@anytime-markdown/editor-core';
+import { ACCENT_COLOR } from '@anytime-markdown/markdown-core';
 import AutoFixHighIcon from '@mui/icons-material/AutoFixHigh';
 import CodeIcon from '@mui/icons-material/Code';
 import ImageIcon from '@mui/icons-material/Image';